+++ CHANGELOG.md
@@ -31,7 +31,6 @@ All notable changes to this project will be documented in this file.
 - When using the Terraform deploy scripts as part of environment setup, workspace GUID values will no longer be wrapped in quotes
 - Storage Type is now required for all Power VS Instances and Volumes regardless of affinity
 - Dropdowns with only one possible item will now be automatically selected within a form
-- Currently, the `eu-es` Madrid region is not supported by Power Cloud API. Until support is added, Power VS Servers are disabled for the `eu-es` region on CRAIG
 
 ### Features
 
@@ -61,6 +60,7 @@ All notable changes to this project will be documented in this file.
 - Fixed an issue causing SCC V2 resources to have an incorrectly formatted resource group in Terraform
 - Fixed an issue causing all Schematics Workspaces to be created in the default resource group even when a resource group was provided
 - Fixed an issue causing a Terraform variable to be generated for Power VS SSH Keys from a data source
+- Fixed an issue with Power API calls within Madrid region to point to `mad` 
 
 ## 1.8.0